third_party/WebKit/LayoutTests/fast/css/image-set-parsing-expected.txt

Test the parsing of the -webkit-image-set function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Single value for background-image : url("#a") 1x
PASS result is "-webkit-image-set(url(\"#a\") 1x)"


Multiple values for background-image : url("#a") 1x, url("#b") 2x
PASS result is "-webkit-image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Multiple values for background-image, out of order : url("#c") 3x, url("#b") 2x, url("#a") 1x
PASS result is "-webkit-image-set(url(\"#c\") 3x, url(\"#b\") 2x, url(\"#a\") 1x)"


Single value for content : url("#a") 1x
PASS result is "-webkit-image-set(url(\"#a\") 1x)"


Multiple values for content : url("#a") 1x, url("#b") 2x
PASS result is "-webkit-image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Single value for border-image : url("#a") 1x
PASS result is "-webkit-image-set(url(\"#a\") 1x)"


Multiple values for border-image : url("#a") 1x, url("#b") 2x
PASS result is "-webkit-image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Single value for -webkit-mask-box-image-source : url("#a") 1x
PASS result is "-webkit-image-set(url(\"#a\") 1x)"


Multiple values for -webkit-mask-box-image-source : url("#a") 1x, url("#b") 2x
PASS result is "-webkit-image-set(url(\"#a\") 1x, url(\"#b\") 2x)"
PASS successfullyParsed is true

TEST COMPLETE
